refactor: use a separate component `PhotoGrid` to render Photo list
